+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Added handwriting font
  - Started saving PDF files so that we can see differences.
  - Changed PageGrouping.appendCell() to take an x-offset as the first parameter (usually set to zero)
+ - Removed LineWrapped.lineHeight because it entirely duplicated .dim.height.
 
 # 2.0.10 - 2018-03-29 "Fill-Rect after many pages"
  - Fixed bug where after a hundred or more pages, the edge of a fill-rect in a PageGrouping could slip between the